LeafGP strategy
