+++ bark/bofire_utils/data_models/strategies/api.py
@@ -1,0 +1,3 @@
+from bark.bofire_utils.data_models.strategies.tree_kernel import (
+    TreeKernelStrategy,  # noqa: F401
+)
