+++ bark/bofire_utils/data_models/surrogates/api.py
@@ -1,6 +1,3 @@
-from bark.bofire_utils.data_models.strategies.tree_kernel import (
-    TreeKernelStrategy,  # noqa: F401
-)
 from bark.bofire_utils.data_models.surrogates.bark import (  # noqa: F401
     BARKPriorSurrogate,
     BARKSurrogate,
